Input field/footer buttonsを統合
